+++ packages/blocks/core/CHANGELOG.md
@@ -117,6 +117,7 @@
 - `Read More` block: support improved.
 - `RSS` block: support improved.
 - `Search` block: support improved.
+- `Site Logo` block: support improved.
 
 
 ### Automated Tests
@@ -171,6 +172,7 @@
 - Updated E2E tests to check support of `Read More` block.
 - Updated E2E tests to check support of `RSS` block.
 - Updated E2E tests to check support of `Search` block.
+- Updated E2E tests to check support of `Site Logo` block.
 
 
 ## 1.3.0 (2025-04-27)